// frontend/src/app/core/models/user-registration.ts
export interface GovernmentUserRegistration {

  

  password: string;

  role: string;

  userType: string;

  personalNumber: string;

  nationalId: string;

  fullName: string;

  phoneNumber: string;

  email: string;

  designation: string;

  county: string;

  subCounty: string;

  division: string;

  location: string;

  subLocation: string;

  village?: string;

}